feat: queryString이 변경될 때 페이지를 렌더링 하도록 설정
