+++ demo/with-cra/package.json
@@ -3,11 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@every-analytics/react-analytics-provider": "file:..",
+    "@every-analytics/react-analytics-provider": "file:../..",
     "@types/node": "^12.0.0",
-    "@types/react": "file:../node_modules/react",
-    "@types/react-dom": "file:../node_modules/react-dom",
-    "react": "file:../node_modules/react",
+    "@types/react": "file:../../node_modules/react",
+    "@types/react-dom": "file:../../node_modules/react-dom",
+    "react": "file:../../node_modules/react",
     "react-dom": "file:../node_modules/react-dom",
     "react-scripts": "4.0.3",
     "typescript": "^4.1.2",
